Use webpack and postcss to build app code
Changes include:
- switching from using CDN's to get the code of dependencies to properly
  installing dependencies using npm
- moving the Vue component's code from being in two large files (one for
  the plugins.html page, one for the processing.html page) to having a
  file for each component, see:
  https://vuejs.org/v2/guide/single-file-components.html
- adding a build process for the javascript code using webpack to get a
  single javascript file to import into each html file rather than
  importing many javascript files
- adding a build process for the css code using postcss

Note that font awesome icons are still used via a CDN.
