+++ web/config/www/.gitignore
@@ -1,5 +1,4 @@
 node_modules/
-package-lock.json
 
-# Compiled JS
-js/
+# Compiled JS and CSS
+dist/
