+++ web/config/www/plugins.html
@@ -1,30 +1,18 @@
 <!DOCTYPE html>
 <html>
-  <head>
-    <meta content="text/html;charset=utf-8" http-equiv="Content-Type">
-    <meta content="utf-8" http-equiv="encoding">
+    <head>
+        <meta charset="utf-8">
+        <meta name="viewport" content="width=device-width,initial-scale=1.0">
 
-    <title>Savu: Plugins</title>
+        <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.6.3/css/all.css" integrity="sha384-UHRtZLI+pbxtHCWp1t77Bi1L4ZtiqrqD80Kn4Z8NTSRyMA2Fd33n5dQ8lWUE00s/" crossorigin="anonymous">
+        <link href="https://fonts.googleapis.com/css?family=Roboto:100,300,400,500,700,900" rel="stylesheet">
+        <link rel="stylesheet" href="dist/tailwind.css" />
+        <link rel="stylesheet" href="css/style.css" />
 
-    <link rel="stylesheet" href="css/style.css" />
-
-    <link href="https://fonts.googleapis.com/css2?family=Lato&display=swap" rel="stylesheet">
-    <link href="https://fonts.googleapis.com/css2?family=Roboto+Slab:wght@300;500&display=swap" rel="stylesheet">
-    <link href="https://unpkg.com/tailwindcss@^1.0/dist/tailwind.min.css" rel="stylesheet">
-
-    <script src="https://cdn.jsdelivr.net/npm/vue/dist/vue.js"></script>
-    <script src="https://unpkg.com/vuex"></script>
-
-    <script src="https://code.jquery.com/jquery-3.3.1.min.js" integrity="sha256-FgpCb/KJQlLNfOu91ta32o/NMZxltwRo8QtmkMRdAu8=" crossorigin="anonymous"></script>
-    <script src="js/api_common.js"></script>
-    <script src="js/api_savu.js"></script>
-    <script src="js/utils.js"></script>
-  </head>
-
-  <body>
-    <div id="vapp">
-        <plugin-page />
-    </div>
-    <script src="src/plugins.js" type="module"></script>
-  </body>
+        <title>Savu: Plugins</title>
+    </head>
+    <body>
+        <div id="vapp"></div>
+        <script src="dist/plugins.bundle.js" type="text/javascript"></script>
+    </body>
 </html>
