+++ web/config/www/postcss.config.js
@@ -1,0 +1,9 @@
+const autoprefixer = require('autoprefixer');
+const tailwindcss = require('tailwindcss');
+
+module.exports = {
+  plugins: [
+    tailwindcss,
+    autoprefixer,
+  ],
+};
