+++ web/config/www/src/api_dawn.js
@@ -1,6 +1,8 @@
+import { plainTextGet } from './api_common.js'
+
 // Gets information about a dataset using the DAWN info API.
 // Accepts a filename and returns on object desribing the dataset.
-function getDataInfo(filename, callback, error) {
+export function getDataInfo(filename, callback, error) {
   plainTextGet("/data/info?path=" + filename, (rawData) => {
     var dataParts = rawData.split("\n");
     var data = {
@@ -17,6 +19,6 @@ function getDataInfo(filename, callback, error) {
 // Accepts a filename and slice parameters as a string.
 // Only two dimensions are allowed to be of size greater than one (no error will
 // be raised but no image will be available at the returned URL).
-function jpegSliceUrl(filename, slice) {
+export function jpegSliceUrl(filename, slice) {
   return "/data/slice?format=JPG&path=" + filename + "&slice=" + slice;
 }
