+++ web/config/www/src/components/JobTabStatusDisplay.vue
@@ -1,0 +1,14 @@
+<template>
+  <div class="w-full flex flex-wrap">
+    <p class="w-full">Status:</p>
+    <span class="w-full">{{ statusText }}</span>
+  </div>
+</template>
+
+<script>
+export default {
+  props: {
+    statusText: String
+  }
+}
+</script>
