+++ web/config/www/src/components/PageTitle.vue
@@ -1,0 +1,9 @@
+<template>
+    <h1 class="mb-2">{{ text }}</h1>
+</template>
+
+<script>
+export default {
+  props: ['text'],
+}
+</script>
