+++ web/config/www/src/components/PluginsPage.vue
@@ -1,0 +1,18 @@
+<template>
+  <div>
+    <page-title text="Plugins" />
+    <plugin-page-grid />
+  </div>
+</template>
+
+<script>
+import PageTitle from './PageTitle.vue'
+import PluginPageGrid from './PluginPageGrid.vue'
+
+export default {
+  components: {
+    'page-title': PageTitle,
+    'plugin-page-grid': PluginPageGrid
+  }
+}
+</script>
