+++ web/config/www/src/components/ProcessingPage.vue
@@ -1,0 +1,18 @@
+<template>
+  <div id="root">
+    <page-title text="Processing" />
+    <processing-page-grid />
+  </div>
+</template>
+
+<script>
+import PageTitle from './PageTitle.vue'
+import ProcessingPageGrid from './ProcessingPageGrid.vue'
+
+export default {
+  components: {
+    'page-title': PageTitle,
+    'processing-page-grid': ProcessingPageGrid
+  }
+}
+</script>
