+++ web/config/www/src/components/TabContent.vue
@@ -1,0 +1,11 @@
+<template>
+  <div>
+    <p>{{ text }}</p>
+  </div>
+</template>
+
+<script>
+export default {
+  props: ['text']
+}
+</script>
